+++ extensions/gitlab/CHANGELOG.md
@@ -5,7 +5,8 @@
 - Add unsupported activity
 - Rename `My Recent Activities` to `Recent Activities`
 - `Recent Activities` got a dropdrown to switch between `My Activities` and `My Projects`
-- Add `Merge Request Menu`
+- Add `Merge Request Menu` command
+- Add `Issues Menu` command
 
 ## [Search Scope] - 2023-04-27